+++ docs/changelog.md
@@ -4,5 +4,6 @@
 
 ### New
 
+* `build-code-coverage-metrics` - covert PHPUnit clover.xml into Markdown and badges
 * `build-contracts` - convert PHPUnit testdox.txt into Markdown
 * `build-php-badges` - get PHP badges from shields.io